pyproject.toml
[tool.poetry]
name = "aws-glue-example-app"
version = "0.1.0"
description = ""
authors = ["Nathan Hadzariga <elnaterator@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "3.3.0"
boto3 = "^1.33.9"
aws-glue-libs = { git = "https://github.com/awslabs/aws-glue-libs.git", branch = 'master' }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
